+++ src/app.module.ts
@@ -3,8 +3,8 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { DatabaseModule } from './database/database.module';
-import { AuthModule } from './auth/auth.module';
-import { RedisModule } from './redis/redis.module';
+import { AuthModule } from './features/auth/auth.module';
+import { RedisModule } from './database/redis/redis.module';
 import * as session from 'express-session';
 import RedisStore from 'connect-redis';
 import { createClient } from 'redis';
