+++ src/database/base.entity.ts
@@ -1,0 +1,16 @@
+import {
+  PrimaryGeneratedColumn,
+  UpdateDateColumn,
+  CreateDateColumn,
+} from 'typeorm';
+
+export class BaseEntity {
+  @PrimaryGeneratedColumn('uuid')
+  id: string;
+
+  @UpdateDateColumn()
+  updated_at: Date;
+
+  @CreateDateColumn()
+  created_at: Date;
+}
