+++ src/features/auth/auth.controller.ts
@@ -1,13 +1,13 @@
 import { Body, Controller, Get, Post, Req, Res } from '@nestjs/common';
-import { RegisterDto } from './dto/register.dto';
+import { RegisterDto } from './auth.dto';
 import { AuthService } from './auth.service';
 
-@Controller()
+@Controller('auth')
 export class AuthContoller {
   constructor(private readonly authService: AuthService) {}
 
   @Get('login')
-  login(@Req() req, @Res() res) {
+  login(@Req() req, @Res() res): void {
     this.authService.login();
     res.send(req.session?.id);
   }
